+++ inst/include/rlasstreamer.h
@@ -59,7 +59,8 @@ class RLASstreamer
     std::vector<unsigned short> R;
     std::vector<unsigned short> G;
     std::vector<unsigned short> B;
-    std::vector< std::vector<double> > ExtraBytes; // extra_byte attributes
+    std::vector<std::vector<int>> ExtraBytes32;     // extra_byte attributes for less than 32 bits types
+    std::vector<std::vector<double>> ExtraBytes64;  // extra_byte attributes for 32 to 64 bits types
 
     LASreadOpener lasreadopener;
     LASwriteOpener laswriteopener;
@@ -87,6 +88,8 @@ class RLASstreamer
     bool p;
     bool rgb;
     std::vector<int> eb; // extra_byte attribute numbers
+    std::vector<int> eb32;
+    std::vector<int> eb64;
 };
 
 #endif //LASSTREAMER_H