+++ NEWS.md
@@ -3,6 +3,7 @@
 #### NEW FEATURES
 
 * new straighforward C++ class `RLASstreamer` to use `rlas`at the C++ level.
+* Support of extra bytes in `readlasdata`
 
 #### BUG FIXES: